refine mcn heads
