+++ configs/simrec_refcoco_scratch.py
@@ -11,12 +11,12 @@ dataset.image_path["refcoco"] = "/home/rentianhe/dataset/rec/images/train2014"
 dataset.mask_path["refcoco"] = "/home/rentianhe/dataset/rec/masks/refcoco"
 
 # # Refine training cfg
-train.output_dir = "./output/sync_with_no_sequential"
+train.output_dir = "./output/test_no_syncbn_one_gpu"
 train.batch_size = 32
 train.save_period = 1
 train.log_period = 10
 train.evaluation.eval_batch_size = 32
-train.sync_bn.enabled = True
+train.sync_bn.enabled = False
 
 # Refine model cfg
 model.visual_backbone.pretrained = True
